fix: 캘린더 CCP에서 비즈니스 로직 분리
- components/Calendar 디렉토리 구조 수정(컴포넌트에서 페이지로 이동)
- ScheduleData 타입 추가
