+++ app/hooks/context/useCalendarContext.ts
@@ -1,11 +1,11 @@
 import { useContext } from 'react';
-import { CalendarContext } from '@/components/Calendar/CalendarPicker';
+import { CalendarContext } from '@/components/Calendar';
 
 const useCalendarContext = () => {
   const context = useContext(CalendarContext);
 
   if (context === undefined) {
-    throw new Error('<Input /> 태그 안에서 사용해주세요.');
+    throw new Error('<Calendar /> 태그 안에서 사용해주세요.');
   }
   return context;
 };
